Flatten mech-core compiler module layout
