+++ src/core/src/compiler/constants.rs
