+++ src/core/src/compiler/context.rs
