+++ src/core/src/compiler/sections.rs
